[scopes] Add ScopeChainEntry implementation for source mapped scopes
This CL adds a new class ScopeWithSourceMappedVariables. It's a
ScopeChainEntry implementation that wraps debugger scopes 1:1 but
replaces variable names in some scopes with their authored name.

This functionality already exists, it's just called directly from the
ScopeChainSidebarPane. Instead, we move this logic into a separate
class.

This allows us to move the "extra synthetic scope properties" business
logic from the view layer, into the "ScopeWithSourceMappedVariables"
class in a follow-up.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Ic30d78e081bd3d2f6637f536746e5136e333d144