[RPP] Make Entry Label to not cover the resizer
Calculate the part of the entry that is overlapping the resizer and adjust the height of the entry wrapper to to be the height of the part of the entry that is visible.

For the entries in the main flamechart that are partly visible, hide the label and label connector.

Video: http://screencast/cast/NDU1OTA4MTY1NzMzNTgwOHw4Y2I1OWQxZi1iMA

Bug: 349530077
Change-Id: I87b392469f2067ef540aeddbff967504679b4beb