[scopes] Move 'exception' and 'return value' properites to model layer
These synthetic properties are currently added on the UI layer to
the scopes view.

This CL moves 2 of the 3 synthetic properties to the model layer. We
add a new 'extraProperties' method on the 'ScopeChainEntry' interface.

Long-term we might move towards adding them to the RemoteObject
directly but that is yet to be determined.

To make this work, the DebuggerModel.Scope needs access to the
current paused info. Instead of retrieving the latest paused info
from the model, we just stash the paused info on each call frame.
This makes some sense as call frames are children of the paused
details hierarchically.

Note that we don't change the @description of the UI strings: This
way we don't need to re-translate them. We'll move the translated
scope description and translated synthetic property names to the
UI layer at some point.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Icac5e2f45c12c458ae621b1f81346391e851b6d4