Fix possible exception when sorting Retainers pane
In some cases, it is possible that the "Restore ignored retainers"
button does not show up after ignoring a retainer. The sequence of steps
in the failure is as follows:

1. HeapSnapshotGridNode.sort() calls populateChildren with toPosition
   greater than 100.
2. populateChildren requests the first 100 items from the worker.
3. The worker responds with fewer than 100 items.
4. populateChildren requests more items, starting at index 100.
5. The worker throws an exception because the start index is invalid.
6. sort() fails, so the remaining steps of
   HeapSnapshotRetainmentDataGrid.dataSourceChanged don't run. One of
   those skipped steps would have shown the "Restore ignored retainers"
   button.

The fix is simple: don't request out-of-bounds data from the worker.
This CL also removes an unclear todo comment from 2012.

Bug: 327337527
Change-Id: I3770c4f7386804039df63f2350ebdf2b15dd466c