RPP: move overlays into timeline/overlays
As we work on the Insights sidebar and need Insights to create Overlays,
it was problematic that the Overlays were not within their own module,
as a component within timeline/component could not depend on the file to
import the types.

This CL makes no functional changes, but moves Overlays into their own
module `timeline/overlays`, which will make it easier to depend on the
types from other parts of the timeline panel.

Bug: 343388196
Change-Id: I8cd586a972b194747d34bdbb8119a01188d4203e