Always include numeric values in heap snapshots
When we added an option to include numeric values in heap snapshots, the
heap snapshot generator was astonishingly slow. At that time, adding
more numeric strings would have a compounding effect on the slowness.
However, that problem was fixed by crrev.com/c/4428811 . Now, I see no
compelling reason to ever omit the numeric values from the heap
snapshot, so this change permanently enables capturing of numeric
values. I did comparisons on several websites and found that this caused
a slowdown of roughly 4% and increased the size of the .heapsnapshot
files by about 2%. I think that the benefit of removing a checkbox from
a complicated tool outweighs those minor performance regressions.

Bug: 40763322
Change-Id: I80567612f90824d5266ac2e27338df19331f75a6