front_end/third_party/lit/README.chromium update Revision
This Revision has been derived using the provided version.

Revisions are now required for dependencies with a git upstream.

This is an initial attempt to uplift the current README metadata.

Respond with Code-Review '-1' if this README file is automatically
updated, you will need to update your automation.

If you need help, or it's not feasible,
file a bug similar to b/352221788 (android_deps).

FAQ:
* Google-internal: go/fix_chromium_metadata
* External: https://docs.google.com/document/d/1pejJ6JSS4lfDfnZ23VfpT1qJ1irDHMBueauCza6tNuo

This CL was uploaded by git cl split.

R=jacktfranklin@chromium.org

Bug: 350851535
Change-Id: I349f95d7ee34d23c5304e16d06818060618b2875