[RPP] Move EntriesFilter to Timeline and check if an entry the `Expand` icon is added to is visible in the timeline
Move EntriesFilter to Timeline because EntriesFilter needs a function that checks if the entry is visible in the Timeline to correctly add the "Expand" button and the `entryIsVisibleInTimeline` function is using a function from EventUICategory, which cannot be imported into the `models`.

Also add a test to cover the case when the next parent is not visible in the timeline.

Bug: 338977625
Change-Id: I52120c97cdeba87231bcdcc2f39a6eff710f62e4