Fix memory leak in ObjectPropertiesSectionsTreeExpandController
This CL fixes a small leak. Currently we use a 'Map' to cache
property paths for tree elements in the object UI (and also the
section map). We never delete from this map so we just keep
accumulating entries. For every pause on the "new tab page" and
expanding the 'globalThis' object this adds 800+ new entries to the
cache map.

This CL replaces 'Map' with 'WeakMap', which fixes the leak (verified
with heapsnapshots).

Drive-by: Move the maps as static private members and use private
field syntax.

R=pfaffe@chromium.org

Bug: None
Change-Id: Ib12795a78ebf05b9812285108ec57d519e0dbba7