[RPP] Fix toggling and overlays for sidebar insights
This sets up a system for knowing what insights are being toggled to
help with expanding & overlay logic.
Insights have their own name, and these
are toggled by the single-navigation-component accordingly.
These are controlled by dispatched events that pass in the
insight+navigationId info so that the correct overlay is displayed.

Bug:344550388
Also-By: aixba@google.com
Change-Id: I3734d3a908a96d3b30f0f598f000821e1a950bda