Update enhanced traces handler logic
We discovered that script with large size could cause issues for enhanced traces while collecting traces, so we updated the traces to split large script into smaller chunks.
https://chromium-review.googlesource.com/c/v8/v8/+/5731354

This patch in dt-fe updates the logic in enhanced traces handler to handle small chunks of large script. Additionally, the handler ensures all the scripts come with auxData if possible as that is critical information for rehydrated session.

Bug: 337909145
Change-Id: I40dd5f56d4b4971847dedf459f736623a6d92280