Add platform-X to `html` not `body`
When trying to use the new `--sys-typescale-X` variables in
designTokens.css, I discovered that they did not have access to the
`--default-font-family` variable.

This is because that variable is defined via the `.platform-X` classes,
which were being added to the `body`. However, the design tokens are all
added to `:root`, which is the `html` element, and therefore are higher
in the tree than the `--default-font-family` variable.

This CL moves the adding of the `platform-X` class onto the `html`
element, which means the `--default-font-family` can be accessed from
`:root`.

Bug: 355622977
Change-Id: I4a4269534dc86b3db2b5dac32cd15d0a8f4897fb