[RPP] Make the Time Range label editable
When we allow to create ranges as annotations, we will save them if the user selects something else in the Flamechart and the range label is not empty.

With this CL, we allow to add a label, but the time range cannot be saved as annotation yet and will disappear when something else is selected.
If the label is empty, do not remove focus from it. In the future, if the label is empty and the focus is removed, we will  remove the whole time range.

Video: http://screencast/cast/NjQ0NjM3Njk5MTg1MDQ5NnxjZmY2ZWRjZS0xNg

Bug: 339802929
Change-Id: I8ca7d74176ff6f55edc0ac24e3fd1598dadf15dd