[l10n] Collect `@meaning` tags in messages too to add explicit meanings
The reason I added this was b/343009012 where they
asked us to provide a `meaning` for the message so that
the TC ID will be different than the other `tab` messages.

As a behavior change, if there is a `@meaning` given to the
message, it won't be replaced with its description when there is a
duplicate message.

Fixed: 343009012
Change-Id: Ib45684aee875b8ddcbc3a9b332006d1d61fe3754