Add checkbox to disable test addresses in the autofill popup
This is disabled by default, in this case the the test addresses option
will not be displayed in the Autofill popup.

Please note that the developer is required to restart devtools for it
to take effect.

demo: https://screencast.googleplex.com/cast/NjI3NjM3MTU4MzA3NDMwNHw1NDU3MzgxMC1jMw
Bug: 40270486
Change-Id: If32621353a7e378b4c558456de9d7150a8c42d06