Add checkbox to disable test addresses in the autofill popup
This is disabled by default, in this case the the test addresses option
will not be displayed in the Autofill popup.

Please note that the developer is required to restart devtools for it
to take effect.

demo: https://screencast.googleplex.com/cast/NTM1MDUwNjQ1NjE1NDExMnw1ZTM2MTE4OS02MA
Bug: 40270486
Change-Id: If32621353a7e378b4c558456de9d7150a8c42d06