[coverage] Use Text object from ContentData
We still keep the contentProvider -> Text map around for now. This is
because for stylesheets we re-request the stylesheet text from the
backend every time otherwise. And then we still wouldn't share the
Text instance as we'd create a new ContentData each time.

At least for Script objects we now share the line number array
across coverage and scopes view.

R=pfaffe@chromium.org

Bug: 356997171
Change-Id: Ieb655b8e1c8952f4b19b5c608b0e00954d9342eb