[RPP] Allow to create multiple Time Range Annotations
If a new time range is selected in the FlameChart and the current time range already has a label, create a new Time Range.

If the Time Range is empty when something else in the Flamechart is selected or Enter is pressed, remove the current Time Range annotations and its overlay.

Video: http://screencast/cast/NTg2MDA5ODM1MzEzNTYxNnxjMzA0MmZkYy1iMA

Bug: 339802929
Change-Id: I89a719283fd755369982ea7f2df95e35100329fe