Add Detached DOM Elements Profiling Type to Memory Panel to get retained JavaScript objects on the page.
Creates a new Detached Elements view that uses the new getDetachedDomNodes CDP command to obtain trees of detached objects that are retained by a JavaScript reference. These trees will be displayed in the view where the specifically retained objects in the detached tree will be marked.

This CL is dependent on this Blink CL
https://chromium-review.googlesource.com/c/chromium/src/+/5634107

Bug: 350519222
Change-Id: I80801d2eacaac8b06cfea256354b3cd6f0de8c77