[RPP]: enable overlay for LCP discovery insight
This contructs the overlay using 1) the candy stripe bar, 2) the error
entry outline, and 3) a singular timespan breakdown. The timespan breakdown
bit is a bit hacky for rendering a singular breakdown for now.
I'm thinking it makes sense to create a seperate overlay type for that,
and so will do this on a follow up.

Bug:351758501
Change-Id: Id2704a85487d6c87b59edf507fba72ed94a76b29