[RPP] zoom into an overlay when an insight's overlay is initially active
This is a small feature to make the insight overlay experience a bit better.
There are cases where the overlay is so tiny that a user would have to
go looking for the overlay. This zooms in a bit when the insight is
initially selected. I'm "zooming" in to where the whole overlay covers
about 50% of the view - but of course this could be changed!

Bug:344550388
Change-Id: Ib531155c85d9a2cba492646844b31439e2fc72e5