RPP: highlight network search results
Currently when the user searches the panel using Ctrl/Cmd-F, we only
search the "main" data provider, which means all network requests are
excluded.

This CL updates the search to include the network track. It required a
few structural changes; rather than store search results as a list of
entryIndexes, we now have to store which provider they came from, so we
know which flame chart to talk to to highlight the right event.

Fixed: 355347482
Change-Id: I30db3496a641e6e56d9c3d3251cb8872509c3eba