RPP: implement new ViewMode for panel
I have been struggling for a while with the logic of how the performance
panel chooses what to render, as we have 3 different states:

1. landing page
2. status pane / "loading" / "processing"
3. viewing a trace

This state was stored across multiple variables and made it very hard to
ensure when we swapped from one state to another that everything was
tidied up. The size of the `#setModel` method showed this well, along
with all the conditionals within it.

This CL creates a new `ViewMode` type and instance property on the
class. We use the new `#changeView` method to set the type of View, and
that takes care of everything. When you are in the `VIEWING_TRACE` mode,
we also store the index of the trace from the model. This ensures you
can never set the mode to `VIEWING_TRACE` without a valid model index.

This then means we can simplify the method that updates the UI for a
trace; it no longer has to deal with the case where it is purposefully
called with `-1` to "reset" the panel and it can deal only with the
positive case of loading a trace.

The motivation here is now that the landing page has the local/field
metrics on, we need a way to return back to the landing page without
loosing all the traces the user may have recorded. That will be done in
a follow-up, but this CL changes the code to make that much easier to
implement.

As a drive-by, I also updated some variables relating to the brick
breaker game ("fixMe" was a confusing name) and moved some private TS
methods into actual private methods.

Bug: 356560849
Change-Id: I0e4e8607ea076998a872571231a479ac77345106