Type host config as a RecursivePartial
This is a defense against changes in the backend implementation of
`getHostConfig` in `devtools_ui_bindings.cc`. In case the shape of the
`HostConfig` object changes, this will now not cause errors in
DevTools anymore, since DevTools is able to handle the absence of any
attribute in `HostConfig`.

Drive-by: disable logging by default, only enable it if `HostConfig`
says so.

Bug: 348136212
Change-Id: I58669154f503dbf28a7e4c44fa35e817e1963df9