[Freestlyer] Add `title` to thoughts so that we can collapse them under this title
Drive-by:
Updated specific constraints from using please to MUST.

Improve success on the z-index question by adding
explicit instructions. (not sure if we should include them though,
this is just for @alexrudenko to see & evaluate the result.)

Maybe, in the future, we can try to include these specific instructions
per question using a system like RAG: we can have embeddings on the
question/domain specific instructions, query them after the question
comes and include the results in the prompt.

Bug: none for now
Change-Id: I2d55ef129fe6d69f62cdf316dbdd775ee0c29bc8