RPP: fix importing multiple traces
In crrev.com/c/5776682 I restructured how traces are loaded, but I made
a mistake with the call to save modifications when we swap traces. My
changes incorrectly lead to the flow being:

1. User imports Trace A
2. User imports Trace B
3.  => we saveModifications(), but we accidentally save A's
    modifications into B, because we have already swapped to Trace B
    being active

This had the result of the timeline data being empty (for reasons that
are not yet entirely clear).

This CL moves the call to store the modifications so it happens before
we swap, ensuring there is no conflict and fixing the issue that would
entirely break the view.

Fixed: 359111053
Change-Id: I4fcebb6d732d1f94367e694ea092fe6f46e4451d