[RPP] Arrow follows the mouse when "Create Entry Link" is selected from the Context Menu
Make EntriesLinkOverlay cover the whole Overlays element and add coordinates for the arrow start and end to the EntriesLinkOverlay. Update those coordinates as the mouse moves over the overlay.

The following CL will add a way to get rid of the arrow and remove the mousemove listener after the entry the arrow connect to is selected.

Video: https://screencast.googleplex.com/cast/NTgzMzY5NzYyNTM3NDcyMHwxOTY1YjkxMC03NQ

Bug: 339802345
Change-Id: If0d18adf70d5990a1da3722be49ed16f915e181c