Heap snapshots: store class names in unused bits of detachedness field
A proposed upcoming change will introduce a new case where the name of
a heap snapshot node doesn't match its class name, and the class name
can't be easily computed from the name. For that change, we'll need a
way to store the class name separately from the name.

With this change, DevTools will store the index of each node's class
name in the previously unused bits of the detachedness field. This
change is intended to be a pure refactor, but there are a few minor
changes that result:

1. The string table is effectively limited to 2**30 entries. I'm pretty
   sure that DevTools would crash due to other limits if a snapshot had
   such a large string table, but I'm mentioning it here anyway for
   completeness.
2. Class names based on node types, such as "RegExp" and "(system)", are
   now included in the string table rather than represented as negative
   values in the classIndex() function.
3. Heap snapshots must include the detachedness field. V8 added this
   field four years ago, so I think we are now beyond the point where
   anybody would expect backward compatibility with the old heap
   snapshot format. (Also, we don't have tests for backward
   compatibility, so it might have already been broken in other ways.)

Bug: 356183770
Change-Id: I75248ca7b73250621120f71de48c9df2913d9b6e