Add "Reveal in Network panel" to Performance tool
When a network request is captured on the Performance timeline, it can be
difficult to reason about why network requests take place. This change
adds the capability to "Reveal in Network panel" if the request is found
within the recent network captures via the Network tool.

A minor refactor was added: Revealers can now optionally accept a property 
bag of options instead of a single `omitFocus` boolean parameter. This is
accomplished by providing an appropriate overload and Revealables can 
optionally define a new method to call. If the method is not implemented
on a Revealable, the old `omitFocus`-only API is called.

Bug: 359279478
Change-Id: I03d108945f576f542ffd034c85a005f873e28561