Fixed up tests
Add "Reveal in Network panel" to Performance tool

When a network request is captured on the Performance timeline, it can be
difficult to reason about why network requests take place. This change
adds the capability to "Reveal in Network panel" if the request is found
within the recent network captures via the Network tool.

Bug: 359279478
Change-Id: I03d108945f576f542ffd034c85a005f873e28561