[RPP]: enable overlay for LCP discovery insight
This constructs the overlay using 1) the candy stripe bar, 2) the error
entry outline, and 3) a singular timespan breakdown. The timespan breakdown
bit is a bit hacky for rendering a singular breakdown for now.
I'm thinking it makes sense to create a separate overlay type for that,
and so will do this on a follow up.

This also fixes a slight bug where the candy stripe bar would follow
when scrolling. When the overlay's event would be hidden, it would
cause the function to end early without correctly updating the style
of the bar.

Bug:351758501
Change-Id: Id2704a85487d6c87b59edf507fba72ed94a76b29