[RPP] Show stack trace for user timings and extension entries
After https://chromium-review.googlesource.com/c/chromium/src/+/5772755
we include the call frame data of performance.mark and measure
invocation in their respective trace events. This CL implements the
frontend side of the feature so that the call site of user timings and
extension entries is visible in the details drawer (and points to the
sources panel if possible).

Bug: 358307050
Change-Id: I79e078dd3c2c2a339a1f7f5fa12b68a320640bd4