[test] Remove unit test relying on focus-visible behavior
This got flaky recently, when another test calling .click() was added.

The reason for that is that focus-visible is based on heuristic, which as implemented in Blink check for the last focus source and whether document saw a native keyboard event.
See https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/css/selector_checker.cc;l=2509

The last focus source is explicitly ignoring focus changes with JS and
dispatching KeyboardEvent from JS doesn't set HadKeyboardEvent() on a
document.

OTOH calling click() in element counts as setting focus with a mouse and
the last focus source is preseved on a document across unit tests.

Finally, the test in question seems to be verifying the browser
behavior more then component implementation and is probably not worth
keeeping given the above.

Bug: none
Change-Id: Ie15147731af93607392aabeeb9b4875fa4a8787b