[RPP Observations] Do not restore interactions log from the buffer
From a user perspective, this means that the interaction log will
always be empty after opening DevTools for the first time.

Due to a limitation in PerformanceObserver, we cannot restore every
interaction that occurred before the observer starts listening. I think it's confusing to only show a subset of interactions, and only
logging interactions that happen while DevTools is open still makes
intuitive sense.

Bug: 334926021
Change-Id: I541251642fb75cbd4c4162ed38019f132e6c54e8