Move ServerTiming implementation to core/platform
The utility parses Server-Timing response headers. We which need such
functionality in the performance panel to show server timings in the
timeline as proposed in go/cpq:server-timings-in-rpp.

So that the trace model in devtools can import it without depending on
the sdk module, the implementation of the util is moved to
core/platform.

Unfortunately, the current implementation of the ServerTiming parser
relies on the i18n pipeline, which we cannot depend on from the
trace/ module nor from the core/platform module. As such, the
implementation is moved without the calls to the i18n pipeline for
UIStrings. Instead, the strings are passed as parameters to the util.
A wrapper is kept in the original place of the util (sdk) which passes
the localized strings, however from the trace engine we use non
localized strings. In the context of the trace engine these strings will
only be console warned, so having them non localized shouldn't be a real
problem.

Drive-by change: support "start" server timing param.
Bug: 340811171
Change-Id: I0c59225e8f4dcbad360d378a9c9d1af73eb57573