Improve names of plain JS objects in heap snapshots
Plain JavaScript objects currently show up in heap snapshots as
'Object'. This CL generates names for them based on the properties they
contain, so that an object's name might look like
'{firstProperty, secondProperty, ..., ninthProperty, tenthProperty}'.
This can help users to tell objects apart more easily when looking at a
big list of them. This CL does not affect the categorization of those
objects in the Summary or Comparison views; they are still in a single
category named 'Object'.

Bug: 356183770
Change-Id: I787ec97b4201f786af38d2534b812d49966794a2