[RPP Observations] Check for iframes on each event
See b/40884049. Accessing `window.top`/`window.self` in an isolated
execution context will create an extra execution context. This adds
extra noise to the execution context dropdown in the console panel so
we should avoid it if we can.

This CL removes the live metrics' usage of `window.top` top detect
iframes, and moves that check to LiveMetrics.ts after events have been
received.

This also renames the live metrics world name to be more friendly as
this will appear in the console panel context dropdown:

https://screenshot.googleplex.com/5F9bcLGAwjmAAzd

Bug: 334926021
Change-Id: I11c8e25bc91e76eb8dfb923547e201f0cb8986e8