[RPP] Create multiple links
Add a variable for a the entries link that is in the process of being created so we can updated it's coordinates on mousemove.

Also allow to create multiple arrows by setting the current link to null in Flamechart when the second entry was selected.

Video: http://screencast/cast/NDY4MTgwNzc5NzY4MjE3Nnw5MDQ3MzY2MS0zNQ

Change-Id: I570ba84f68792a4313a813a595f30ae29daa8ab3