Allow searching for synthetic names in heap snapshots
Currently, if you press ctrl-f in a heap snapshot and type
"concatenated", you'll find a bunch of search results that don't
actually include the word "concatenated", but are of type "(concatenated
string)". This result breaks the principle of least surprise: the search
box should find results according to the text displayed on the screen
for each object. The problem is that for some objects (for concatenated
strings since forever, and for plain JS Objects since recently), the
name displayed on the screen is synthesized on demand and doesn't match
the object's name in the original snapshot data. Thus, the search
feature should look for matches in these synthetic names rather than the
raw names from the snapshot.

This change brings up another problem: if we iterate every node in the
snapshot calling name() on them, then we'll fill up the lazyStringCache
with synthetic names, at a large memory cost. I think that this cache is
not actually gaining us much, since the frontend usually only requests
100 names at a time, and the worker thread can easily synthesize 100
names without any noticeable delay. To avoid high memory usage, this CL
removes the lazyStringCache.

Bug: 356183770
Change-Id: I339251a96ee45bd3e8e1dbb150a56163dd17d6f1