[NavigatorView] Add unit tests for domain node display name
Follow up to: https://crrev.com/c/5787610

This also updates the logic to consider if the default context is not
the first in the list and if the default context has a context name. Not
certain if either of these cases will ever happen in practice but it's
good to cover them regardless.

Bug: 349392953
Change-Id: Iffa85dffc445e02c30c37202ebda23f43a99f894