[RPP] Support markdown in insights
This makes it simpler to localize strings with links or code tags.

Using markdown inside an i18n string is new for CDT, but not for
the i18n library DevTools uses (which was taken from Lighthouse, and
markdown is used liberally there). There's nothing special to do to
use Markdown, all the markdown-specific bits are transformed to
placeholders for the i18n pipeline like so:

  "panels/timeline/components/insights/Viewport.ts | description": {
    "message": "A viewport meta element not only optimizes your app for mobile screen sizes, but also $LINK_START_0$prevents a 300 millisecond delay to user input$LINK_END_0$.",
    "description": "Text to tell the user how a viewport meta element can improve performance.",
    "placeholders": {
      "LINK_START_0": {
        "content": "["
      },
      "LINK_END_0": {
        "content": "](https://developer.chrome.com/blog/300ms-tap-delay-gone-away/)"
      }
    }
  },

This patch utilizes the existing MarkdownView component, but with a
helper that allows rendering given just the markdown string.

Bug: None
Change-Id: I2cc1af7aed2918cde090c69d9ef7c8a3a1fd4622