[RPP Observations] Separate metric card component into new file
This is just a helpful refactor to reduce file sizes as we start adding
more environment recomendations. Any shared code between the two files
was moved to Util.ts.

The test cases should be split up, but I'm leaving that for a follow-up
to ensure that just moving code around doesn't break anything.

Bug: 313906438
Change-Id: Icbe6597154b43d31acef4ec286d50e9b9679e475