Devtools: Enable `Capability.Network` for Worklet targets
This is a followup to a previous patchset:
https://chromium-review.googlesource.com/c/chromium/src/+/5726981
which added support for intercepting and instrumenting network requests
made from worklets running on a worker thread such as `AudioWorklet`s.

In order for that code to get activated, the `Network` capability needs
to be activated for that debugger target.  I've tested that once that
capability is enabled, requests made from within audio worklet
processors are successfully picked up by devtools (they currently don't
appear).

Bug: 40686236
Bug: 40592236
Change-Id: I81ef14b9e6b6161885a5ee7cacddd83b259008f0