Correctly compute dominators in the Memory tool
This change doesn't yet implement a fix! It just adds some very slow
verification code. It would be unwise to run this verification code on
large snapshots; the snapshot I've been testing with is under 40,000
nodes (from the chrome://inspect page).

I've verified the verifier manually: I inspected a few of the incorrect
relationships that it flagged in the Retainers pane, and found that
there were indeed other paths to the root not involving the supposed
dominator node.

Bug: 361372448
Change-Id: Ia3d881265a7dba12c785eb4d954a52b4fef2fc29