Correctly compute dominators in the Memory tool
The function HeapSnapshot.buildDominatorTree often terminates too early,
as described in the linked bug. It should not terminate until it has
visited every node and found no changes for any of them.

However, continuing the algorithm until it terminates takes longer, so
this change adds a new heuristic in an effort to counteract that
slowdown: if a node has lots of predecessors and didn't change the last
time it was visited, avoid adding it to the "affected" list. On a
particularly troublesome snapshot where the current incorrect algorithm
takes 45 seconds, and a simple fix increases that time to 87 seconds,
this heuristic brings the time back down to 46 seconds.

This is not the right long-term strategy; 46 seconds is still way too
long! We should replace this algorithm with the standard Lengauer-Tarjan
algorithm, which has much more appealing asymptotic complexity.

Bug: 361372448
Change-Id: Ia3d881265a7dba12c785eb4d954a52b4fef2fc29